# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5ebbbc9ebbf7210dd3bd64a3
  title: meetup
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /product:
    post:
      x-skaffolder-id: 5ebbbd37bbf7210dd3bd653e
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5ebbbd2abbf7210dd3bd6531
      x-skaffolder-resource: Product
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Product
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category:
                  type: string
                product_name:
                  type: string
                quantity:
                  type: number
    get:
      x-skaffolder-id: 5ebbbd38bbf7210dd3bd6544
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5ebbbd2abbf7210dd3bd6531
      x-skaffolder-resource: Product
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Product
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  "/product/{id}":
    delete:
      x-skaffolder-id: 5ebbbd38bbf7210dd3bd6545
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ebbbd2abbf7210dd3bd6531
      x-skaffolder-resource: Product
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Product
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ebbbd37bbf7210dd3bd6540
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ebbbd2abbf7210dd3bd6531
      x-skaffolder-resource: Product
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Product
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    post:
      x-skaffolder-id: 5ebbbd37bbf7210dd3bd653f
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ebbbd2abbf7210dd3bd6531
      x-skaffolder-resource: Product
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Product
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
...
  /user:
    post:
      x-skaffolder-id: 5ebbbc9ebbf7210dd3bd64b6
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5ebbbc9ebbf7210dd3bd64ad
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5ebbbc9ebbf7210dd3bd64b1
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5ebbbc9ebbf7210dd3bd64ad
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 5ebbbc9ebbf7210dd3bd64b0
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ebbbc9ebbf7210dd3bd64ad
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ebbbc9ebbf7210dd3bd64b7
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ebbbc9ebbf7210dd3bd64ad
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5ebbbc9ebbf7210dd3bd64b3
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ebbbc9ebbf7210dd3bd64ad
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      x-skaffolder-id: 5ebbbd2abbf7210dd3bd6531
      x-skaffolder-id-db: 5ebbbc9ebbf7210dd3bd64a4
      x-skaffolder-id-entity: 5ebbbd2abbf7210dd3bd652d
      x-skaffolder-url: /product
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        category:
          type: string
          x-skaffolder-id-attr: 5ebbbd2abbf7210dd3bd652f
          x-skaffolder-type: String
        product_name:
          type: string
          x-skaffolder-id-attr: 5ebbbd2abbf7210dd3bd652e
          x-skaffolder-type: String
          x-skaffolder-required: true
          x-skaffolder-unique: true
        quantity:
          type: number
          x-skaffolder-id-attr: 5ebbbd2abbf7210dd3bd6530
          x-skaffolder-type: Number
      x-skaffolder-relations: 
      required:
        - _id
        - product_name
    User:
      x-skaffolder-id: 5ebbbc9ebbf7210dd3bd64ad
      x-skaffolder-id-db: 5ebbbc9ebbf7210dd3bd64a4
      x-skaffolder-id-entity: 5ebbbc9ebbf7210dd3bd64a5
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5ebbbc9ebbf7210dd3bd64a8
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5ebbbc9ebbf7210dd3bd64a9
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5ebbbc9ebbf7210dd3bd64a7
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5ebbbc9ebbf7210dd3bd64ab
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5ebbbc9ebbf7210dd3bd64aa
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5ebbbc9ebbf7210dd3bd64a6
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...